Moved button thread to its own file
